Restore missing quotas in new graph-based collector
Quotas were missing because ems_ref was being set to the tenant_id
which is not unique. Changed ems_ref to [tenant_id, quota name] to
match what was being used as lookup value to find_or_build.

Also removed collection of quotas that we are not modeling:
metadata items, injected files, server groups, and rbac policies.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1512641